+++ sheerwater_benchmarking/forecasts/__init__.py
@@ -1,7 +1,7 @@
 """Forecasting models for the Sheerwater benchmarking project."""
 from .salient import salient
 from .abc import perpp
-from .ecmwf_er import ecmwf_abc_wb, ecmwf_ifs_er, ecmwf_ifs_er_debiased
+from .ecmwf_er import ecmwf_ifs_er, ecmwf_ifs_er_debiased
 
 # Use __all__ to define what is part of the public API.
 __all__ = ["ecmwf_abc_wb", "salient", "ecmwf_ifs_er_debiased", "ecmwf_ifs_er", "perpp"]
