+++ sheerwater_benchmarking/reanalysis/__init__.py
@@ -1,9 +1,8 @@
 """Data functions for all parts of the data pipeline."""
-from .era5 import era5_agg, era5, era5_daily, era5_rolled
+from .era5 import era5, era5_daily, era5_rolled
 
 # Use __all__ to define what is part of the public API.
 __all__ = [
-    era5_agg,
     era5,
     era5_rolled,
     era5_daily
