+++ sheerwater_benchmarking/reanalysis/era5.py
@@ -1,117 +1,12 @@
-"""Fetches ERA5 data from the CDS API."""
-import dask
-import cdsapi
-import os
+"""Fetches ERA5 data from the Google ARCO Store."""
 import xarray as xr
 import dateparser
 import numpy as np
 from datetime import datetime, timedelta
 
 from sheerwater_benchmarking.utils import (dask_remote, cacheable,
-                                           cdsapi_secret,
-                                           get_grid, get_variable,
-                                           apply_mask, clip_region, roll_and_agg, lon_base_change,
-                                           regrid, get_anomalies)
-
-
-@cacheable(data_type='array', cache_args=['year', 'variable', 'grid'])
-def single_era5(year, variable, grid="global1_5"):
-    """Fetches a single year of hourly ERA5 data.
-
-    Args:
-        year (str): The year to fetch data for.
-        variable (str): The weather variable to fetch.
-        grid (str): The grid resolution to fetch the data at. One of:
-            - global1_5: 1.5 degree
-    """
-    variable = get_variable(variable, 'era5')
-
-    times = ['00:00', '01:00', '02:00',
-             '03:00', '04:00', '05:00',
-             '06:00', '07:00', '08:00',
-             '09:00', '10:00', '11:00',
-             '12:00', '13:00', '14:00',
-             '15:00', '16:00', '17:00',
-             '18:00', '19:00', '20:00',
-             '21:00', '22:00', '23:00']
-    days = [str(i) for i in range(1, 32)]
-    months = ["01", "02", "03", "04", "05", "06",
-              "07", "08", "09", "10", "11", "12"]
-
-    _, _, grid_size = get_grid(grid)
-
-    url, key = cdsapi_secret()
-    c = cdsapi.Client(url=url, key=key)
-
-    os.makedirs('./temp', exist_ok=True)
-    path = f"./temp/{variable}.{year}.{grid}.nc"
-
-    print(f"Fetching data {variable} data for year {year}, months {
-          months}, days {days}, and times {times} at grid {grid}.")
-
-    c.retrieve('reanalysis-era5-single-levels',
-               {
-                   'product_type': 'reanalysis',
-                   'variable': variable,
-                   'year': year,
-                   'month': months,
-                   'day': days,
-                   'time': times,
-                   'format': 'netcdf',
-                   'grid': [str(grid_size), str(grid_size)],
-               },
-               path)
-
-    # Read the data and return individual datasets
-    ds = xr.open_dataset(path)
-    ds = ds.compute()
-    os.remove(path)
-    return ds
-
-
-@cacheable(data_type='array',
-           cache_args=['year', 'variable', 'grid'],
-           chunking={"lat": 121, "lon": 240, "time": 1000})
-def single_era5_cleaned(year, variable, grid="global1_5"):
-    """Fetches a single year of ERA5 data and cleans it up.
-
-    Handles the two different formats of ERA5 data.
-    Has the same interface as single_era5.
-    """
-    ds = single_era5(year, variable, grid)
-
-    # Handle second data format
-    if "valid_time" in ds.coords:
-        ds = ds.rename({'valid_time': 'time'})
-    if "expver" in ds.coords:
-        ds = ds.drop_vars("expver")
-    if "number" in ds.coords:
-        ds = ds.drop_vars("number")
-
-    return ds
-
-
-@dask_remote
-def era5_cds(start_time, end_time, variable, grid="global1_5"):
-    """Aggregate yearly ERA5 data into a single dataset."""
-    # Read and combine all the data into an array
-    first_year = dateparser.parse(start_time).year
-    last_year = dateparser.parse(end_time).year
-    years = [str(y) for y in range(first_year, last_year+1)]
-
-    datasets = []
-    for year in years:
-        ds = dask.delayed(single_era5_cleaned)(
-            year, variable, grid, filepath_only=True)
-        datasets.append(ds)
-
-    ds = dask.compute(*datasets)
-    ds = xr.open_mfdataset(ds,
-                           engine='zarr',
-                           parallel=True,
-                           chunks={'lat': 121, 'lon': 240, 'time': 1000}
-                           )
-    return ds
+                                           get_variable, apply_mask, clip_region,
+                                           roll_and_agg, lon_base_change, regrid)
 
 
 @dask_remote
@@ -129,6 +24,8 @@ def era5_raw(start_time, end_time, variable, grid="global0_25"):  # noqa ARG001
                       chunks={'time': 50, 'latitude': 721, 'longitude': 1440})
 
     # Select the right variable
+    if variable in ['tmax2m', 'tmin2m']:
+        var = 'tmp2m'  # Compute min and max daily temperatures from 2m temperature
     var = get_variable(variable, 'era5')
     ds = ds[var].to_dataset()
 
@@ -168,15 +65,30 @@ def era5_daily(start_time, end_time, variable, grid="global1_5"):
     # Convert to base180 longitude
     ds = lon_base_change(ds, to_base="base180")
 
+    K_const = 273.15
     if variable == 'tmp2m':
-        ds[variable] = ds[variable] - 273.15
+        ds[variable] = ds[variable] - K_const
         ds.attrs.update(units='C')
         ds = ds.resample(time='D').mean(dim='time')
+    if variable == 'tmax2m':
+        ds[variable] = ds[variable] - K_const
+        ds.attrs.update(units='C')
+        ds = ds.resample(time='D').max(dim='time')
+    if variable == 'tmin2m':
+        ds[variable] = ds[variable] - K_const
+        ds.attrs.update(units='C')
+        ds = ds.resample(time='D').min(dim='time')
     elif variable == 'precip':
         ds[variable] = ds[variable] * 1000.0
         ds.attrs.update(units='mm')
         ds = ds.resample(time='D').sum(dim='time')
+        # Can't have precip less than zero (there are some very small negative values)
         ds = np.maximum(ds, 0)
+    elif variable == 'ssrd':
+        ds = ds.resample(time='D').sum(dim='time')
+        ds = np.maximum(ds, 0)
+    else:
+        raise ValueError(f"Variable {variable} not implemented.")
     return ds
 
 
@@ -236,61 +148,6 @@ def era5_rolled(start_time, end_time, variable, agg=14, grid="global1_5"):
     agg_fn = "sum" if variable == "precip" else "mean"
     ds = roll_and_agg(ds, agg=agg, agg_col="time", agg_fn=agg_fn)
 
-    return ds
-
-
-@dask_remote
-@cacheable(data_type='array',
-           timeseries='time',
-           cache=True,
-           cache_args=['variable', 'agg', 'anom', 'clim_params', 'grid', 'mask', 'region'],
-           chunking={"lat": 121, "lon": 240, "time": 1000},
-           chunk_by_arg={
-               'grid': {
-                   'global0_25': {"lat": 721, "lon": 1440, 'time': 30}
-               }
-           },
-           auto_rechunk=False)
-def era5_agg(start_time, end_time, variable, agg=14,
-             anom=False, clim_params={'first_year': 1991, 'last_year': 2020},
-             grid="global1_5", mask="lsm", region='global'):
-    """Fetches ground truth data from ERA5 and applies aggregation and masking.
-
-    Specialized function for the ABC model.
-
-    Args:
-        start_time (str): The start date to fetch data for.
-        end_time (str): The end date to fetch.
-        variable (str): The weather variable to fetch.
-        agg (str): The aggregation period to use, in days
-        anom (bool): Whether to return the climatological anomaly
-        clim_params (dict): Parameters for the climatology anomaly calculation,
-            passed directly to climatology function. If anom is False,
-            these parameters are ignored.
-        grid (str): The grid resolution to fetch the data at. One of:
-            - global1_5: 1.5 degree global grid
-        mask (str): The mask to apply to the data. One of:
-            - lsm: Land-sea mask
-            - None: No mask
-        region (str): The region to clip the data to. One of:
-            - global: Global region
-            - africa: Africa region
-    """
-    # Get ERA5 on the corresponding global grid
-    ds = era5_rolled(start_time, end_time, variable, agg=agg, grid=grid)
-
-    if anom:
-        # Import here to avoid circular dependency
-        from sheerwater_benchmarking.baselines import climatology_raw
-        # Get the climatology on the same grid
-        clim = climatology_raw(variable, **clim_params, grid=grid)
-        # Get the anomalies
-        ds = get_anomalies(ds, clim, var=variable)
-
-    # Apply masking
-    ds = apply_mask(ds, mask, var=variable, grid=grid)
-    # Clip to specified region
-    ds = clip_region(ds, region=region)
     return ds
